move internal tools containers to containers. fix some more
